requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
flatbuffers==2.0
gast==0.4.0
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
idna==3.10
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
numpy==2.1.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyodbc==5.2.0
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
shiboken6==6.8.0.2
six==1.16.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.16.0
zxing-cpp==2.2.0
